+++ CHANGELOG.md
@@ -49,6 +49,10 @@ Added / Fixed / changed for 2.2.0 alpha 1
 ### Known Issues
   - This is the first release of the v2.2 code
 
+Added / Fixed / changed for 2.1.4
+-----------------------
+### This release only fixes bugs found in 2.1.3, everything else is destined for the v2.2 major version update
+
 Added / Fixed / changed for 2.1.3
 -----------------------
 ### This release only fixes bugs found in 2.1.2, everything else is destined for the v2.2 major version update